refactoring: created error-message entity slice, moved error from shared. Created lib segment and moved errorTypePredicates and getErrorMessage functions into it.
