+++ src/entities/error-message/index.ts
@@ -1,0 +1,3 @@
+export * from './model/types';
+
+export { getErrorMessage } from './lib/getErrorMessage';
