+++ src/entities/error-message/model/types.ts
