--- src/shared/error/index.ts
@@ -1,3 +1,0 @@
-export * from './model/types';
-
-export { getErrorMessage } from './model/getErrorMessage';
